Move ExternalWorkspace class
